+++ talking_keyboard/audio.py
@@ -5,7 +5,7 @@ import logging
 import os
 import time
 
-import alsaaudio
+# import alsaaudio
 import pygame
 import requests
 from gtts import gTTS
@@ -15,17 +15,17 @@ from const import COMMON_WORDS_FILE, MP3_DIR
 _LOGGER = logging.getLogger(__name__)
 
 
-class AlsaMixer:
-    def __init__(self, mixer_name="PCM", cardindex=1):
-        self.mixer = alsaaudio.Mixer('Headphone', cardindex=3)
-        self.volume = self.getvolume()
-    
-    def getvolume(self):
-        return self.mixer.getvolume()[0]
+# class AlsaMixer:
+#     def __init__(self, mixer_name="PCM", cardindex=1):
+#         self.mixer = alsaaudio.Mixer("Headphone", cardindex=3)
+#         self.volume = self.getvolume()
 
-    def set_volume(self, vol):
-        self.mixer.setvolume(vol)
-        _LOGGER.info("Volume set to %d", vol)
+#     def getvolume(self):
+#         return self.mixer.getvolume()[0]
+
+#     def set_volume(self, vol):
+#         self.mixer.setvolume(vol)
+#         _LOGGER.info("Volume set to %d", vol)
 
 
 class GoogleTTS:
@@ -67,10 +67,6 @@ class PygameMP3Player:
         self.load_common_words()
 
         try:
-            # Check if audio devices are available
-            if len(alsaaudio.cards()) == 0:
-                raise Exception("No ALSA audio devices found.")
-
             pygame.init()
             self.player = pygame.mixer
             self.player.init()
@@ -80,7 +76,6 @@ class PygameMP3Player:
         except Exception as e:
             _LOGGER.error(f"General error initializing audio: {e}")
             raise
-
 
     def preload_sound(self, text):
         filename = os.path.join(MP3_DIR, f"{text}.mp3")
