+++ talking_keyboard/main.py
@@ -1,13 +1,10 @@
 import argparse
 import logging
-import os
 import subprocess
-import sys
 import threading
 import time
 
-from const import COMMON_LETTERS, KEY_MAP
-from keyboard import Keyboard
+from loop import Loop
 
 
 def parse_arguments():
@@ -35,12 +32,8 @@ logging.basicConfig(
 )
 
 _LOGGER = logging.getLogger(__name__)
-
-# if os.geteuid() != 0:
-#     _LOGGER.error("This script must be run with sudo privileges.")
-#     sys.exit(1)
-
 _LOGGER.info("Starting up with log level %d", numeric_level)
+
 
 def check_internet_connection():
     try:
@@ -49,62 +42,52 @@ def check_internet_connection():
     except subprocess.CalledProcessError:
         return False
 
+
 def update_wpa_supplicant(ssid, psk):
     wpa_supplicant_path = "/etc/wpa_supplicant/wpa_supplicant.conf"
 
     with open(wpa_supplicant_path, "a") as f:
-        f.write(f"\nnetwork={{\nssid=\"{ssid}\"\npsk=\"{psk}\"\n}}\n")
+        f.write(f'\nnetwork={{\nssid="{ssid}"\npsk="{psk}"\n}}\n')
 
     subprocess.call(["sudo", "systemctl", "daemon-reload"])
     subprocess.call(["sudo", "systemctl", "restart", "dhcpcd"])
+
 
 def get_user_input(prompt):
     user_input = ""
 
     while True:
         char = input()
-        if char == '\n':
+        if char == "\n":
             break
-        elif char == '\b':
+        elif char == "\b":
             user_input = user_input[:-1]
         elif char:
             user_input += char
 
     return user_input
 
+
 if __name__ == "__main__":
-        _LOGGER.info("Starting talking keyboard")
+    _LOGGER.info("Starting talking keyboard")
 
+    wifi = check_internet_connection()
+    # Check internet connection
+    while not wifi:
+        time.sleep(2)
+        ssid = get_user_input("wifi SSID:")
+        psk = get_user_input("wifi PSK:")
+        update_wpa_supplicant(ssid, psk)
+        time.sleep(5)
         wifi = check_internet_connection()
-        # Check internet connection
-        while not wifi:
-            time.sleep(2)
-            ssid = get_user_input("wifi SSID:")
-            psk = get_user_input("wifi PSK:")
-            update_wpa_supplicant(ssid, psk)
-            time.sleep(5)
-            wifi = check_internet_connection()
-            time.sleep(2)
-            
-        keyboard = Keyboard()
+        time.sleep(2)
 
-        _LOGGER.info("Preloading common letters")
-        for letter in COMMON_LETTERS:
-            if f" {letter} " not in keyboard.player.generated_words:
-                _LOGGER.info("    Preloading letter: %s", letter)
-                keyboard.player.preload_sound(f" {letter} ")
-        keyboard.player.save_common_words()
+    loop = Loop()
+    loop.preload()
 
-        _LOGGER.info("Preloaded words are:")
-        for word in keyboard.player.generated_words.keys():
-            _LOGGER.info("    %s", word)
+    save_thread = threading.Thread(
+        target=loop.player.periodic_save, args=(300,), daemon=True
+    )
+    save_thread.start()
 
-        keyboard.word = "Bonjour, bienvenue sur le clavier parlant."
-        keyboard.process_letter("\n", False)
-
-        save_thread = threading.Thread(
-            target=keyboard.player.periodic_save, args=(300,), daemon=True
-        )
-        save_thread.start()
-
-        keyboard.loop()
+    loop.loop()
